requirements/base.txt
loguru
celery
environs
redis
flower==1.2.0
boto3
httpx
singleton_decorator
torch==2.3.0
sentence-transformers==2.7.0
huggingface_hub